add functionality to rate a book

+++ Domain/Rating.cs
@@ -2,7 +2,7 @@ namespace Domain
 {
     public class Rating : BaseEntity<int>
     {
-        public decimal Score { get; set; }
+        public int Score { get; set; }
         public Book Book { get; set; }
     }
 }
